Change clickable Cards to be based off a single implementation and provide an analytics event
