+++ aries-site/src/components/home/Featured.js
@@ -1,7 +1,6 @@
 import React from 'react';
 
 import { Box, Grid, PageContent, Paragraph, Stack, Text } from 'grommet';
-import Link from 'next/link';
 import { ContentPreviewCard } from '../cards';
 import { nameToPath } from '../../utils';
 import { featured } from '../../data';
@@ -10,26 +9,23 @@ const FeaturedLayout = ({ ...rest }) => (
   <PageContent {...rest}>
     <Grid columns={{ count: 'fit', size: 'small' }} gap="large">
       {featured.map(({ name, description, icon, url }) => (
-        <Link href={url || nameToPath(name)} passHref key={name}>
-          <ContentPreviewCard
-            as="a"
-            style={{ textDecoration: 'none' }}
-            pad="medium"
+        <ContentPreviewCard
+          key={name}
+          href={url || nameToPath(name)}
+          pad="medium">
+          <Box width="100%" round="xsmall">
+            {icon}
+          </Box>
+          <Text
+            weight="bold"
+            size="xlarge"
+            color="text-strong"
+            margin={{ top: 'small' }}
           >
-            <Box width="100%" round="xsmall">
-              {icon}
-            </Box>
-            <Text
-              weight="bold"
-              size="xlarge"
-              color="text-strong"
-              margin={{ top: 'small' }}
-            >
-              {name}
-            </Text>
-            <Paragraph size="small">{description}</Paragraph>
-          </ContentPreviewCard>
-        </Link>
+            {name}
+          </Text>
+          <Paragraph size="small">{description}</Paragraph>
+        </ContentPreviewCard>
       ))}
     </Grid>
   </PageContent>
